Now building to HTML files etc. and Caddy as dev webserver.
The build phase actually builds to HTML files etc. now and [Caddy](https://caddyserver.com), now the development webserver of the project,  uses those files because I can't for the life of it figure out how to configure the Vite webserver properly (with Content Security Policy, Permissions Policy and reversy proxy to the database API).